+++ src/pages/Admin/components/RightSection/ListPostContentBox.tsx
@@ -1,12 +1,15 @@
 import styled from 'styled-components';
 
 const ListPostContentsBox = ({ data, openModal }) => {
+  const date = new Date(data.updated_at);
+  const convertedDate = date.toISOString().substring(0, 10);
+
   return (
     <ListContents onClick={openModal}>
       <UserName>{data.user_name}</UserName>
       <PostContent>{data.content}</PostContent>
       <UserMbti>{data.user_mbti}</UserMbti>
-      <UserSignDate>{data.updated_at}</UserSignDate>
+      <UserSignDate>{convertedDate}</UserSignDate>
     </ListContents>
   );
 };
